pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-based-file-organizer"
version = "0.1.0"
description = "Automated file organization system using content analysis and LLM-based naming"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "watchdog>=3.0.0",
    "PyPDF2>=3.0.0",
    "boto3>=1.28.0",
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov",
    "black",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
